+++ Project/SecretDetection/SecretsAnalysis/DictionaryDetector.cs
@@ -10,10 +10,10 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 namespace Project.SecretDetection.SecretsAnalysis{
     public class DictionaryDetector : Detector
     {
-        public int score = 0;
-        public override int detect(string secret)
+        public float score = 0.0F;
+        public override float detect(string secret)
         {
-            return 0; // Brug evt. den her: https://github.com/loresoft/NetSpell
+            return 0.0F; // Brug evt. den her: https://github.com/loresoft/NetSpell
                       // Skal researches først
                       // Eksempel brug af NetSpell: https://stackoverflow.com/questions/38416265/c-sharp-checking-if-a-word-is-in-an-english-dictionary
